Conflicts:
	G8Taquin/src/g8taquin/Grille.java
Méthode : forçage du int

+++ G8Taquin/src/g8taquin/Case.java
@@ -28,7 +28,7 @@ public class Case {
 
     //renvoie l'attribut x 
     public int getCoordx(){    
-       return this.x; 
+       return (int) this.x; 
     }
     
     //renvoie l'attribut y
